move constant as suggested
